# chenoj-frontend/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "接口文档",
    "contact": {}
  },
  "host": "localhost:8121",
  "basePath": "/api",
...
  "paths": {
...
    "/api/user/add": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "addUser",
        "operationId": "addUserUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "userAddRequest",
            "description": "userAddRequest",
            "required": true,
            "schema": {
              "originalRef": "UserAddRequest",
              "$ref": "#/definitions/UserAddRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«long»",
              "$ref": "#/definitions/BaseResponse«long»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«long»",
              "$ref": "#/definitions/BaseResponse«long»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/user/delete": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "deleteUser",
        "operationId": "deleteUserUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "deleteRequest",
            "description": "deleteRequest",
            "required": true,
            "schema": {
              "originalRef": "DeleteRequest",
              "$ref": "#/definitions/DeleteRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«boolean»",
              "$ref": "#/definitions/BaseResponse«boolean»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«boolean»",
              "$ref": "#/definitions/BaseResponse«boolean»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/user/get": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "getUserById",
        "operationId": "getUserByIdUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "id",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«User»",
              "$ref": "#/definitions/BaseResponse«User»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«User»",
              "$ref": "#/definitions/BaseResponse«User»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/user/get/login": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "getLoginUser",
        "operationId": "getLoginUserUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«LoginUserVO»",
              "$ref": "#/definitions/BaseResponse«LoginUserVO»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«LoginUserVO»",
              "$ref": "#/definitions/BaseResponse«LoginUserVO»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/user/get/vo": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "getUserVOById",
        "operationId": "getUserVOByIdUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "id",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«UserVO»",
              "$ref": "#/definitions/BaseResponse«UserVO»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«UserVO»",
              "$ref": "#/definitions/BaseResponse«UserVO»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/user/list/page": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "listUserByPage",
        "operationId": "listUserByPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "userQueryRequest",
            "description": "userQueryRequest",
            "required": true,
            "schema": {
              "originalRef": "UserQueryRequest",
              "$ref": "#/definitions/UserQueryRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«Page«User»»",
              "$ref": "#/definitions/BaseResponse«Page«User»»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«Page«User»»",
              "$ref": "#/definitions/BaseResponse«Page«User»»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/user/list/page/vo": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "listUserVOByPage",
        "operationId": "listUserVOByPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "userQueryRequest",
            "description": "userQueryRequest",
            "required": true,
            "schema": {
              "originalRef": "UserQueryRequest",
              "$ref": "#/definitions/UserQueryRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«Page«UserVO»»",
              "$ref": "#/definitions/BaseResponse«Page«UserVO»»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«Page«UserVO»»",
              "$ref": "#/definitions/BaseResponse«Page«UserVO»»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/user/list/pagewr": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "listUserByPageWhenRegister",
        "operationId": "listUserByPageWhenRegisterUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "userQueryRequest",
            "description": "userQueryRequest",
            "required": true,
            "schema": {
              "originalRef": "UserQueryRequest",
              "$ref": "#/definitions/UserQueryRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«Page«User»»",
              "$ref": "#/definitions/BaseResponse«Page«User»»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«Page«User»»",
              "$ref": "#/definitions/BaseResponse«Page«User»»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/user/login": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "userLogin",
        "operationId": "userLoginUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "userLoginRequest",
            "description": "userLoginRequest",
            "required": true,
            "schema": {
              "originalRef": "UserLoginRequest",
              "$ref": "#/definitions/UserLoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«LoginUserVO»",
              "$ref": "#/definitions/BaseResponse«LoginUserVO»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«LoginUserVO»",
              "$ref": "#/definitions/BaseResponse«LoginUserVO»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/user/login/wx_open": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "userLoginByWxOpen",
        "operationId": "userLoginByWxOpenUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "description": "code",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«LoginUserVO»",
              "$ref": "#/definitions/BaseResponse«LoginUserVO»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«LoginUserVO»",
              "$ref": "#/definitions/BaseResponse«LoginUserVO»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/user/logout": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "userLogout",
        "operationId": "userLogoutUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«boolean»",
              "$ref": "#/definitions/BaseResponse«boolean»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«boolean»",
              "$ref": "#/definitions/BaseResponse«boolean»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/user/register": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "userRegister",
        "operationId": "userRegisterUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "userRegisterRequest",
            "description": "userRegisterRequest",
            "required": true,
            "schema": {
              "originalRef": "UserRegisterRequest",
              "$ref": "#/definitions/UserRegisterRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«long»",
              "$ref": "#/definitions/BaseResponse«long»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«long»",
              "$ref": "#/definitions/BaseResponse«long»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/user/update": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "updateUser",
        "operationId": "updateUserUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "userUpdateRequest",
            "description": "userUpdateRequest",
            "required": true,
            "schema": {
              "originalRef": "UserUpdateRequest",
              "$ref": "#/definitions/UserUpdateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«boolean»",
              "$ref": "#/definitions/BaseResponse«boolean»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«boolean»",
              "$ref": "#/definitions/BaseResponse«boolean»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
...
  },
  "definitions": {
...
    "BaseResponse«LoginUserVO»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "originalRef": "LoginUserVO",
          "$ref": "#/definitions/LoginUserVO"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "BaseResponse«LoginUserVO»"
    },
...
    "BaseResponse«UserVO»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "originalRef": "UserVO",
          "$ref": "#/definitions/UserVO"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "BaseResponse«UserVO»"
    },
    "BaseResponse«User»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "originalRef": "User",
          "$ref": "#/definitions/User"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "BaseResponse«User»"
    },
    "BaseResponse«boolean»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "BaseResponse«boolean»"
    },
...
    "BaseResponse«long»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "integer",
          "format": "int64"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "BaseResponse«long»"
    },
...
    "DeleteRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "DeleteRequest"
    },
...
    "UserAddRequest": {
      "type": "object",
      "properties": {
        "userAccount": {
          "type": "string"
        },
        "userAvatar": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "userRole": {
          "type": "string"
        }
      },
      "title": "UserAddRequest"
    },
...
    "UserQueryRequest": {
      "type": "object",
      "properties": {
        "current": {
          "type": "integer",
          "format": "int32"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "mpOpenId": {
          "type": "string"
        },
        "pageSize": {
          "type": "integer",
          "format": "int32"
        },
        "sortField": {
          "type": "string"
        },
        "sortOrder": {
          "type": "string"
        },
        "unionId": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "userProfile": {
          "type": "string"
        },
        "userRole": {
          "type": "string"
        }
      },
      "title": "UserQueryRequest"
    },
...
  },
...
}